DBZ-7223 Add the MongoDB sink connector * remove yet unused, commented code * move MongoDbSinkConnector class to root level package * address PR feedback about the container NETWORK change in MongoDbDatabaseProvider and possible side-effects
closes https://issues.redhat.com/browse/DBZ-7223
